Start covering native integration with tests
